requirements.txt
pillow
matplotlib
numpy
scikit-image
opencv-python
scipy